# Discussion: speed up refresh for EDDTableFromMultidimNcFiles