Hi @BobSimons - I'm opening a PR here mostly because it's the easiest way to discuss code changes -- but if you'd rather discuss this on the mailing list or as a github issue, I'm happy to move the conversation there instead.

**Background**

I'm trying to set up an ERDDAP server with datasets based on very large netcdf files. On the upper end, we're talking 50GB datasets that are composed of individual nc files that are 500MB each. 

I know that ERDDAP will take a while for the initial dataset load, since it's loading in the netcdf data and building its own table representation, and that's fine. However, after that initial load, I need to reload to be relatively fast, so that I don't have to wait hours or days for a reload if the server goes down and I need to restart. But this isn't what I was seeing. Here's an example:

* Initial load:
 `*** EDDTableFromFiles test-dataset_NC constructor finished. TIME=8701ms`
* Subsequent load (no modifications to the file at all in between; just restart ERDDAP):
 `*** EDDTableFromFiles test-dataset_NC constructor finished. TIME=4427ms`

This is only a 2x speed improvement -- and that's about what I was seeing for other test netcdf datasets, ranging in file size.

**Changes/Questions**

So I started debugging to see what was going on. There were three things I found that I wanted to ask you about. 

1. **`makeExpected`: update `lastMod` check**

    When reloading, `EDDTableFromFiles::makeExpected` checks to see if the file has changed since the last time it loaded. I found that there were cases where the file mod time was *less than* the last mod time that erddap had stored
    * example: `1588697736060 -- ftLastMod.get(f)` and `1588697736000 -- lastMod` -- only off by 60ms

    So i switched this check from `ftLastMod.get(f) == lastMod` to `ftLastMod.get(f) < lastMod`. For one of my test cases, this made it so that I went from around 5.5s for a reload of a 20MB nc file, to a .7s reload time, which makes sense -- if it thought the file had been modified, then it would do a full reload which is expensive.

    While I'm not sure why the file table and file last mod times are different at all, this seems like a reasonable fix? Thoughts?

2. **`EDDTableFromMultidimNcFiles`: `mustGetData` flag?**

   There is a flag, `mustGetData` that is passed to `lowGetSourceDataFromFile`. This flag is set to false during reload, because ERDDAP should not have to load up all the data at this point (after all, it's already checked that the file hasn't been modified since it last looked at it). However, that flag wasn't getting honored by `EDDTableFromMultidimNcFiles` (it isn't passed into `readMultidimNc`), so it was *loading up the data every single time*.  

    I started digging into this, but `readMultidimNc` is very complex so I didn't get very far. Plus my update (3) below made it so that `getSourceDataFromFile` is never called anyway. But I'm mostly curious why `EDDTableFromMultidimNcFiles` doesn't honor the `mustGetData` param -- it seems like updating it to do so would drastically improve reload speeds for netcdf file datasets.

3. **`makeExpected`: skip `getSourceDataFromFile`**
 
    After looking at the `mustGetData` issue, I was wondering why ERDDAP has to touch the nc file at all, if it hasn't been modified? In other words: in `EDDTableFromFiles.makeExpected`, if the file hasn't changed and it isn't the same size, then just completely skip loading up the `Table` for that file. 

    If I remove that code block (see changes) then it goes from 0.7s --> 0.2s for a reload. So overall, that's 5.5s-->0.2s, an over 90% decrease in time spent reloading. And I've run this branch with dozens of large datasets and saw a similar speedup, with no functionality problems that I could find.

    That said, I'm sure there is some reason that code block is in there? Are there other modifications that ERDDAP is looking for?

Thanks for reading, I look forward to hearing your responses.