@BobSimons 2020-07-20
I understand the desire for a faster reload in quickRestart mode. That's the point of quickRestart mode.

To answer your questions:

1) The problem with switching to ftLastMod.get(f) < lastMod is: You're thinking "The file is 60ms older than expected. That's a quirk of the operating system, so let's ignore changes where the file is older than expected." But I'm thinking that an older file might be much older (e.g., 60 days older) because it is an older version of the file and therefore needs to be read and processed. So I am reluctant to make that change to the code.  

You could then say "well, then let's make ERRDAP ignore changes of <1 s. Perhaps. But It is possible that some real time dataset is changed very frequently, so it could be a different version of the file. So I am even reluctant to make that change.

I have no explanation of why the OS says the lastModified time is slightly different. That's disappointing. 

2)  The parent class EDDTableFromFiles deals with a wide range of subclasses, so the methods that the subclasses have to handle (notably lowGetSourceDataFromFile) sometimes include parameters that are requests, not obligations.  (It's actually more complicated than that, e.g., the whole CONSTRAIN_NO/PARTIAL/YES system.)  mustGetData (note that it isn't just called "getData") is thus stating a requirement that the implementation must get the data (otherwise it is optional). 

So why doesn't EDDTableFromMultidimNcFiles pay attention and use that parameter? You basically answered that: "readMultidimNc is very complex". I didn't take the time to add that option to the reader by adding to the complexity of an already complex method. 

3) The whole point of makeExpected is to get the expected values of certain vital metadata attributes from one of the source files that is known to be valid.  This information is used elsewhere to validate new or changed files to ensure that the variables have the same (expected) values of scale_factor, add_offset, missing_value, etc. so we know that aggregating all these files is a valid thing to do.  Most of the time, opening one file from a collection doesn't take long. But here it does. So we can't just remove that code.

There is a possible way around this: the dataset could store this information in a file in the dataset's info directory. But that adds considerable complexity to the process. So I am reluctant to do that.

Another option is: we could make makeExpected lazy. Even that might be somewhat complicated since I think the information is used in other places (e.g., updateEveryNMillis). I will investigate.

---
I feel bad. You did a lot of work to do the analysis and make the changes and I rejected your 3 suggested changes and left you without a solution. I'll say: the situation I envisioned for these datasets was: the aggregation consists of 1 to a million files, each of which can be read fairly quickly. When that is true, the time needed for makeExpected is minimal because it only ever opens 1 file, and thus quickRestart finishes quickly if no files have changed.  And I envisioned that restarting ERDDAP was a rare event. Your situation is different because reading one file takes significant time and you might have lots of datasets like this.

I'll mention an obvious solution: make the data files somewhat smaller. That has a few advantages. 
1) It solves this problem.
2) It makes it more likely that ERDDAP can read less data when responding to requests simply by quickly rejecting more of these smaller files. (But at the expense of sometimes having to open up lots more files. It's a trade-off.)
But I'm guessing you don't want to make that change. Fair enough.

I will try to think of other solutions. The ones I have thought of so far just add more complexity to an already fairly complex system. If you have other possible solutions, please let me know.



Another option is