@jessicaaustin 2020-07-20
@BobSimons Thanks for the quick reply and thorough response!

> I feel bad. You did a lot of work to do the analysis and make the changes and I rejected your 3 suggested changes and left you without a solution.

You shouldn't feel bad; part of this was me validating I had set up my config properly. There are countless ways to configure an ERDDAP dataset and I wanted to understand what it was doing under the hood so I could make the right decisions about how to structure my datasets. With datasets this large, a small tweak can make all the difference. (Plus I was able to set up a unit test to try out different changes quickly, so the debugging was not as bad as you might think) 

I will say the only thing I found "surprising" was that even if my files did not change whatsoever, ERDDAP was still loading up all the data for the first file -- in other words that EDDTableFromMultidimNcFiles reads the data not just the metadata. 

> I'll say: the situation I envisioned for these datasets was: the aggregation consists of 1 to a million files, each of which can be read fairly quickly. When that is true, the time needed for makeExpected is minimal because it only ever opens 1 file, and thus quickRestart finishes quickly if no files have changed. 

Yes, this makes sense and is what I expected after looking at the code. So even if my dataset is 50GB, if it's broken up into very small files then the reload should be quick because it only has to look at one file to check the metadata against what's stored in the file table. I will think about this; there may be something I can do here, at least to make the first file smaller. 

> And I envisioned that restarting ERDDAP was a rare event. Your situation is different because reading one file takes significant time and you might have lots of datasets like this.

It should still be a rare event, but even if it's rare I need the reload to happen in a reasonable amount of time. For example, the [IOOS ESM ERDDAP](http://erddap.sensors.axds.co/erddap/index.html) takes about an hour to fully reload after restart, which means we do major config changes or upgrades on off hours, but they are still do-able.

> Another option is: we could make makeExpected lazy. Even that might be somewhat complicated since I think the information is used in other places (e.g., updateEveryNMillis). I will investigate.

I'm interested to hear where this goes. Would you consider adding a configuration flag that would make `makeExpected` more or less aggressive about how it does its updates? This could affect the `lastMod` check as well. So for example I would set that flag to "lazy" (if file not changed since last update, do nothing) for datasets with large files that I know change very infrequently, and "aggressive" (current behavior) for datasets that are updated very frequently and I want to make sure ERDDAP stays on top of those changes. 

> Another option is

Did you mean to have more here? It's cut off.