@BobSimons 2020-07-21
Yes, the last "Another option is" was a typo.

I now suspect the best option is to make the code not get the data when mustGetData is false.  I'm super busy right now and the current release is super late, so I will probably put off working on this until the next release.