@jessicaaustin 2020-07-21
Okay, that makes sense. I think as the size of datasets scale, this change will help more and more users, not just our particular use case here.

In the meantime, if I can't split into smaller files then I'll just run a fork with the `lastMod` change, and I can revert that once the netcdf mustGetData updates are in the main trunk a few releases from now.

Thanks!