@shane-axiom 2022-10-09
> the best option is to make the [`EDDTableFromMultidimNcFiles` `lowGetSourceDataFromFile`] code not get the data when mustGetData is false

For reference, Bob implemented this change here, first released in v2.10:

https://github.com/BobSimons/erddap/commit/e55556f18d39d5ae3f4ff96748fec2ebd42ace63#diff-174e5ec59fb0f843a0e59a3ad65f2640ad372f2f226e2db95fb048d35dd49d35